fix: apply lucide-react
