--- src/app/_components/Icon/AddIcon.tsx
@@ -1,27 +1,0 @@
-import React from 'react';
-
-interface AddIconProps extends React.SVGProps<SVGSVGElement>{
-    size?: number;
-}
-
-export const AddIcon: React.FC<AddIconProps> = ({ size, height, width, ...props }) => (
-    <svg
-        version="1.0"
-        xmlns="http://www.w3.org/2000/svg"
-        height={size || height || 20}
-        width={size || width || 20}
-        viewBox="0 0 512.000000 512.000000"
-        preserveAspectRatio="xMidYMid meet"
-        {...props}
-    >
-        <g
-            transform="translate(0.000000,512.000000) scale(0.100000,-0.100000)"
-            fill="currentColor"
-            stroke="none"
-        >
-            <path d="M119 4542 c-70 -37 -111 -100 -117 -180 -3 -48 1 -66 25 -112 18 -34 45 -66 72 -85 l43 -30 293 -3 293 -3 10 -27 c83 -231 876 -2341 891 -2370 28 -54 102 -117 168 -142 54 -20 66 -20 1281 -18 l1227 3 54 30 c56 31 110 86 135 139 29 61 619 1931 623 1977 10 94 -66 202 -160 228 -31 9 -433 11 -1548 9 l-1506 -3 -48 -30 c-36 -22 -57 -45 -77 -84 -61 -116 -21 -241 97 -304 l40 -22 1338 -3 c1212 -2 1338 -4 1333 -18 -5 -13 -446 -1414 -462 -1466 -5 -17 -64 -18 -1063 -18 l-1057 0 -459 1221 c-497 1320 -480 1279 -569 1320 -38 17 -74 19 -422 19 l-382 0 -53 -28z" />
-            <path d="M1923 1256 c-136 -44 -230 -153 -253 -290 -25 -155 55 -308 198 -378 60 -30 75 -33 157 -33 78 0 98 4 153 29 274 127 274 528 0 651 -75 34 -188 43 -255 21z" />
-            <path d="M3925 1257 c-184 -62 -290 -232 -255 -409 23 -121 96 -215 203 -264 54 -25 74 -29 152 -29 82 0 97 3 157 33 272 133 265 524 -11 650 -67 31 -183 39 -246 19z" />
-        </g>
-  </svg>
-);
