+++ src/app/_components/buttons/floatingButton.tsx
@@ -1,28 +1,16 @@
-import React, { useEffect, useState } from 'react';
-import IconButton from './IconButton';
-import { Badge } from '@nextui-org/react';
-import { getLocalStorageItemCount } from '@/app/_managers/localStorageManager';
-import { addLocalStorageListener } from '@/app/_managers/eventListenerManager';
+import React, { useEffect, useState } from "react";
+import { Badge, Card } from "@nextui-org/react";
+import { getLocalStorageItemCount } from "@/app/_managers/localStorageManager";
+import { addLocalStorageListener } from "@/app/_managers/eventListenerManager";
 
 interface FloatingButtonProps {
-  color: 'primary' | 'secondary' | 'success' | 'warning' | 'danger';
-  variant?: 'solid' | 'faded' | 'bordered' | 'light' | 'shadow';
-  ariaLabel: string;
-  hovermsg: string;
-  icon: React.ReactNode;
-  onPress: () => void;
+    color: "primary" | "secondary" | "success" | "warning" | "danger";
+    icon: React.ReactNode;
+    onPress: () => void;
 }
 
-const FloatingButton: React.FC<FloatingButtonProps> = ({
-  color,
-  variant = 'solid',
-  ariaLabel,
-  hovermsg,
-  icon,
-  onPress,
-}) => {
-  
-  const [data, setData] = useState<number>(0);
+const FloatingButton: React.FC<FloatingButtonProps> = ({ color, icon, onPress }) => {
+    const [data, setData] = useState<number>(0);
 
     useEffect(() => {
         // 초기 데이터 로드
@@ -36,26 +24,19 @@ const FloatingButton: React.FC<FloatingButtonProps> = ({
         const unsubscribe = addLocalStorageListener<number>("clickedItemData", fetchData);
 
         return () => {
-          unsubscribe();
+            unsubscribe();
         };
     }, []);
 
-  
-  return (
-    <div className="fixed bottom-10 right-20 z-50">
-        <Badge content={data} shape='circle' color='danger'>
-            <IconButton
-                color={color}
-                variant={variant}
-                ariaLabel={ariaLabel}
-                hovermsg={hovermsg}
-                icon={icon}
-                onPress={onPress}
-                size='lg'
-            />
-        </Badge>
-    </div>
-  );
+    return (
+        <div className="fixed bottom-4 right-4 lg:bottom-16 lg:right-20 z-50">
+            <Badge size="lg" content={data} shape="circle" color={color}>
+                <Card shadow="lg" radius="lg" isPressable onPress={onPress} isHoverable>
+                    {icon}
+                </Card>
+            </Badge>
+        </div>
+    );
 };
 
 export default FloatingButton;
